Move the part that instantiating the python unittest testcases to loader
